requirements.txt
flask
sounddevice
assist
vosk
openai
pygame
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
azure-identity 
requests
twilio
